change primary key and add user_id

+++ memgpt/metadata.py
@@ -604,9 +604,13 @@ class MetadataStore:
 
     @enforce_types
     # def list_tools(self, user_id: uuid.UUID) -> List[ToolModel]: # TODO: add when users can creat tools
-    def list_tools(self) -> List[ToolModel]:
+    def list_tools(self, user_id: Optional[uuid.UUID] = None) -> List[ToolModel]:
         with self.session_maker() as session:
-            results = session.query(ToolModel).all()
+            if user_id:
+                results = session.query(ToolModel).filter(ToolModel.user_id == user_id).all()
+            else:
+                # return all created tools
+                results = session.query(ToolModel).all()
             return results
 
     @enforce_types
@@ -677,10 +681,10 @@ class MetadataStore:
             return results[0].to_record()
 
     @enforce_types
-    def get_tool(self, tool_name: str) -> Optional[ToolModel]:
+    def get_tool(self, tool_name: str, user_id: uuid.UUID) -> Optional[ToolModel]:
         # TODO: add user_id when tools can eventually be added by users
         with self.session_maker() as session:
-            results = session.query(ToolModel).filter(ToolModel.name == tool_name).all()
+            results = session.query(ToolModel).filter(ToolModel.name == tool_name).filter(ToolModel.user_id == user_id).all()
             if len(results) == 0:
                 return None
             assert len(results) == 1, f"Expected 1 result, got {len(results)}"
@@ -752,6 +756,8 @@ class MetadataStore:
     @enforce_types
     def add_tool(self, tool: ToolModel):
         with self.session_maker() as session:
+            if session.query(ToolModel).filter(ToolModel.name == tool.name).filter(ToolModel.user_id == tool.user_id).count() > 0:
+                raise ValueError(f"Tool with name {tool.name} already exists for user_id {tool.user_id}")
             session.add(tool)
             session.commit()
 
@@ -811,9 +817,9 @@ class MetadataStore:
             session.commit()
 
     @enforce_types
-    def delete_tool(self, name: str):
+    def delete_tool(self, name: str, user_id: uuid.UUID):
         with self.session_maker() as session:
-            session.query(ToolModel).filter(ToolModel.name == name).delete()
+            session.query(ToolModel).filter(ToolModel.name == name).filter(ToolModel.user_id == user_id).delete()
             session.commit()
 
     # job related functions
